+++ CHANGES.rst
@@ -268,7 +268,14 @@ csv_tools
 ---------
 
 cube_build
-----------
+---------
+- Removed spaxel.py and replace class with set of arrays [#2472] 
+- reworked in mapping of the detector pixel to the sky spaxel so that consistent
+  code can be used for both MIRI and NIRSPEC data [#2472] 
+- Removed some loops in cube_cloud.py for finding which pixels fall in roi 
+  of spaxels [#2472] 
+- In a test with MIRI data there was a 13% improvement in the speed of making IFUcubes. In the
+  NIRSPEC case there was a 40% improvment in the speed of creating IFUCubes.  
 
 - Fixed bug in cube_build.blot_images that was failing for  NIRSPEC IFU images
   with the slide position defined in the WCS [#2345]